+++ libs/eavmlib/src/network.erl
@@ -27,7 +27,8 @@
     wait_for_ap/0, wait_for_ap/1, wait_for_ap/2,
     sta_rssi/0,
     sta_disconnect/0,
-    sta_connect/0, sta_connect/1
+    sta_connect/0, sta_connect/1,
+    sta_status/0
 ]).
 -export([start/1, start_link/1, stop/0]).
 -export([
@@ -171,6 +172,8 @@
 -type network_config() :: [sta_config() | ap_config() | sntp_config() | mdns_config()].
 
 -type db() :: integer().
+-type sta_status() ::
+    associated | connected | connecting | degraded | disconnected | disconnecting | inactive.
 
 -record(state, {
     config :: network_config(),
@@ -323,7 +326,7 @@ start_link(Config) ->
 %%-----------------------------------------------------------------------------
 -spec sta_disconnect() -> ok | {error, Reason :: term()}.
 sta_disconnect() ->
-    gen_server:call(?SERVER, halt_sta).
+    gen_server:call(?SERVER, halt_sta, 65000).
 
 %%-----------------------------------------------------------------------------
 %% @param   Config The new station mode network configuration
@@ -338,7 +341,7 @@ sta_disconnect() ->
 -spec sta_connect(Config :: network_config() | [sta_config_property()]) ->
     ok | {error, Reason :: term()}.
 sta_connect(Config) ->
-    gen_server:call(?SERVER, {connect, Config}).
+    gen_server:call(?SERVER, {connect, Config}, 65000).
 
 %%-----------------------------------------------------------------------------
 %% @returns ok, if the network interface was started, or {error, Reason} if
@@ -384,6 +387,24 @@ sta_rssi() ->
             end
     end.
 
+%%-----------------------------------------------------------------------------
+%% @returns ConnectionState :: sta_status().
+%%
+%% @doc Get the connection status of the sta interface.
+%%
+%% Results will be one of: `associated', `connected', `connecting', `degraded',
+%% `disconnected', `disconnecting', or `inactive'. The state `associated' indicates
+%% that the station is connected to an access point, but does not yet have an IP address.
+%% A status of `degraded' indicates that the connection has experienced at least one
+%% beacon timeout event during the current connection session. This does not necessarily
+%% mean the connection is still in a poor state, but it might be helpful diagnosing
+%% problems with networked applications.
+%% @end
+%%-----------------------------------------------------------------------------
+-spec sta_status() -> Status :: sta_status().
+sta_status() ->
+    gen_server:call(?SERVER, sta_status).
+
 %%
 %% gen_server callbacks
 %%
@@ -392,7 +413,20 @@ sta_rssi() ->
 init(Config) ->
     Port = get_port(),
     Ref = make_ref(),
-    {ok, #state{config = Config, port = Port, ref = Ref}, {continue, start_port}}.
+    Status =
+        case proplists:get_value(sta, Config) of
+            undefined ->
+                inactive;
+            STA ->
+                case proplists:get_value(managed, STA, false) of
+                    false ->
+                        connecting;
+                    true ->
+                        disconnected
+                end
+        end,
+    {ok, #state{config = Config, port = Port, ref = Ref, sta_state = Status},
+        {continue, start_port}}.
 
 %% @hidden
 handle_continue(start_port, #state{config = Config, port = Port, ref = Ref} = State) ->
@@ -407,17 +441,19 @@ handle_continue(start_port, #state{config = Config, port = Port, ref = Ref} = St
 %% @hidden
 handle_call(halt_sta, _From, #state{ref = Ref} = State) ->
     network_port ! {self(), Ref, halt_sta},
-    wait_halt_sta_reply(Ref, State);
+    wait_halt_sta_reply(Ref, State#state{sta_state = disconnecting});
 handle_call(connect, _From, #state{config = Config, ref = Ref} = State) ->
     network_port ! {self(), Ref, {connect, Config}},
-    wait_connect_reply(Ref, Config, State);
+    wait_connect_reply(Ref, Config, State#state{sta_state = connecting});
 handle_call({connect, Config}, _From, #state{config = OldConfig, ref = Ref} = State) ->
     case update_config(OldConfig, Config) of
         {error, Reason} ->
             {reply, {error, Reason}, State};
         NewConfig ->
             network_port ! {self(), Ref, {connect, NewConfig}},
-            wait_connect_reply(Ref, NewConfig, State#state{config = NewConfig})
+            wait_connect_reply(Ref, NewConfig, State#state{
+                sta_state = connecting, config = NewConfig
+            })
     end;
 handle_call(sta_status, _From, State) ->
     {reply, State#state.sta_state, State};
@@ -431,16 +467,16 @@ handle_cast(_Msg, State) ->
 %% @hidden
 handle_info({Ref, sta_connected} = _Msg, #state{ref = Ref, config = Config} = State) ->
     maybe_sta_connected_callback(Config),
-    {noreply, State};
+    {noreply, State#state{sta_state = associated}};
 handle_info({Ref, sta_beacon_timeout} = _Msg, #state{ref = Ref, config = Config} = State) ->
     maybe_sta_beacon_timeout_callback(Config),
-    {noreply, State};
+    {noreply, State#state{sta_state = degraded}};
 handle_info({Ref, sta_disconnected} = _Msg, #state{ref = Ref, config = Config} = State) ->
     maybe_sta_disconnected_callback(Config),
-    {noreply, State};
+    {noreply, State#state{sta_state = disconnected, sta_ip_info = undefined}};
 handle_info({Ref, {sta_got_ip, IpInfo}} = _Msg, #state{ref = Ref, config = Config} = State0) ->
     maybe_sta_got_ip_callback(Config, IpInfo),
-    State1 = State0#state{sta_ip_info = IpInfo},
+    State1 = State0#state{sta_ip_info = IpInfo, sta_state = connected},
     State2 = maybe_start_mdns(State1),
     {noreply, State2};
 handle_info({Ref, ap_started} = _Msg, #state{ref = Ref, config = Config} = State) ->
@@ -479,20 +515,22 @@ terminate(_Reason, State) ->
 wait_connect_reply(Ref, NewConfig, State) ->
     receive
         {Ref, ok} ->
-            {reply, ok, State#state{ref = Ref, config = NewConfig}};
+            {reply, ok, State#state{config = NewConfig}};
         {Ref, {error, _Reason} = ER} ->
-            {reply, ER, State}
+            {reply, ER, State#state{sta_state = disconnected}}
+    after 60000 ->
+        {reply, {error, timeout}, State#state{sta_state = disconnected}}
     end.
 
 %% @private
 wait_halt_sta_reply(Ref, State) ->
     receive
         {Ref, ok} ->
-            {reply, ok, State};
+            {reply, ok, State#state{sta_state = disconnected}};
         {Ref, {error, _Reason} = Error} ->
             {reply, Error, State}
     after 60000 ->
-        {reply, {error, timeout}, State#state{sta_state = degraded}}
+        {reply, {error, timeout}, State}
     end.
 
 wait_for_port_close(PortMonitor, Port) ->
@@ -521,8 +559,8 @@ maybe_sta_beacon_timeout_callback(Config) ->
 
 %% @private
 maybe_sta_disconnected_callback(Config) ->
-    maybe_callback0(
-        disconnected, proplists:get_value(sta, Config, fun sta_disconnected_default_callback/0)
+    maybe_callback2(
+        disconnected, proplists:get_value(sta, Config), fun sta_disconnected_default_callback/0
     ).
 
 %% @private
@@ -601,6 +639,21 @@ maybe_callback1({Key, Arg} = Msg, Config) ->
     end.
 
 %% @private
+maybe_callback2(_Key, undefined, _Default) ->
+    ok;
+maybe_callback2(Key, Config, Default) ->
+    case proplists:get_value(Key, Config) of
+        undefined when is_function(Default) ->
+            spawn(fun() -> Default() end);
+        Pid when is_pid(Pid) ->
+            Pid ! Key;
+        Fun when is_function(Fun) ->
+            spawn(fun() -> Fun() end);
+        _ ->
+            ok
+    end.
+
+%% @private
 -spec get_port() -> port().
 get_port() ->
     case whereis(network_port) of
@@ -656,14 +709,24 @@ update_config(OldConfig, NewConfig) ->
                 Ap -> {ap, Ap}
             end,
         MDNS =
-            case proplists:get_value(mdns, OldConfig) of
-                undefined -> [];
-                Mdns -> {mdns, Mdns}
+            case
+                update_opts(
+                    proplists:get_value(mdns, OldConfig, []),
+                    proplists:get_value(mdns, NewConfig, [])
+                )
+            of
+                [] -> [];
+                MdnsCfg -> {mdns, MdnsCfg}
             end,
         SNTP =
-            case proplists:get_value(sntp, OldConfig) of
-                undefined -> [];
-                Sntp -> {sntp, Sntp}
+            case
+                update_opts(
+                    proplists:get_value(sntp, OldConfig, []),
+                    proplists:get_value(sntp, NewConfig, [])
+                )
+            of
+                [] -> [];
+                NewSntp -> {sntp, NewSntp}
             end,
         lists:flatten([STA, AP, MDNS, SNTP])
     catch